got rid of Conf class
