+++ srcs/main/Webserv.cpp
@@ -25,7 +25,7 @@ Webserv& Webserv::operator=(const Webserv &rhs)
     if(this != &rhs)
     {
         this->env = rhs.env;
-        this->conf = rhs.conf;
+        this->confs = rhs.confs;
     }
     return(*this);
 }
@@ -49,6 +49,67 @@ char** Webserv::getEnv()
     return(this->env_char);
 }
 
+bool checkNumberBrackets(const vec_string &split)
+{
+    size_t size = split.size();
+    size_t count = 0;
+    for (size_t i = 0; i < size; i++)
+    {
+        if(split[i] == "{")
+        {
+            count++;
+        }
+        else if(split[i] == "}")
+        {
+            if(count == 0 )
+                return 1;
+            count--;
+        }
+    }
+    return 0;
+    
+}
+
+void Webserv::parse(vec_string split)
+{
+    int check = 0;
+    size_t size = split.size();
+    // for(size_t jt = 0; jt < size; jt++)
+    // {
+    //     std::cout << split[jt] << std::endl;
+    // }
+    if(checkNumberBrackets(split))
+        return(errorParsing("Issue with the file, uneven number of {}"));
+    for(size_t i = 0; i < size; i++)
+    {
+        if(split[i] == "server")
+        {
+            i++;
+            try
+            {
+                ServerConf newConf = parser(split, i, size);
+                vec_string name = newConf.getServerNames();
+                this->confs.push_back(std::make_pair(name, newConf));
+                check++;
+            }
+            catch(const std::exception& e)
+            {
+                writeInsideLog(e, errorParsing);
+                // throw;
+            }
+        }
+    }
+    // std::cout << "Test: " << this->confs.size() << std::endl;
+    if(check)
+    {
+        std::cout << "Coucou" << std::endl;
+        printConfig(this->confs);  
+    }
+    else
+        throw std::invalid_argument("Webserv: Error: No configuration found");
+
+}
+
 void Webserv::parseConfig(const std::string &conf)
 {
     std::ifstream config;
@@ -62,5 +123,5 @@ void Webserv::parseConfig(const std::string &conf)
     strm << config.rdbuf();
     std::string str = strm.str();
     config.close();
-    this->conf.parse(tokenizer(str, " \n\t\r\b\v\f", "{};"));
+    this->parse(tokenizer(str, " \n\t\r\b\v\f", "{};"));
 }
